refactor: split notification views
