+++ gsl_notification/templates/gsl_notification/choose_upload_document_type.html
@@ -1,0 +1,55 @@
+{% extends "gsl_notification/tab_simulation_projet/tab_notifications.html" %}
+{% load static gsl_filters %}
+{% block notification_content %}
+    <a class="fr-btn fr-btn--icon-left fr-btn--tertiary fr-icon-arrow-left-s-line fr-mb-4w"
+       href="{% url 'gsl_notification:documents' programmation_projet.id %}">
+        Annuler la génération
+    </a>
+
+    <div class="arrete-container" data-controller="choose-document-type">
+        <h2 class="fr-mb-3w">
+            Générer un document
+        </h2>
+
+        <div class="fr-radio-group fr-mb-2w"
+             {% if programmation_projet.arrete_et_lettre_signes %}aria-describedby="tooltip-import-arrete-et-lettre-signes"{% endif %}>
+            <input type="radio"
+                   name="type"
+                   value="arrete-et-lettre-signes"
+                   dsfr="dsfr"
+                   id="arrete-et-lettre-signes"
+                   {% if programmation_projet.arrete_et_lettre_signes %}disabled {% endif %}
+                   data-choose-document-type-target="radio"
+                   data-action="click->choose-document-type#setSelectedValue">
+            <label class="fr-label" for="arrete-et-lettre-signes">
+                Arrêté et lettre de notification signés
+            </label>
+            {% if programmation_projet.arrete_et_lettre_signes %}
+                <span class="fr-tooltip fr-placement"
+                      id="tooltip-import-arrete-et-lettre-signes"
+                      role="tooltip">
+                    Ce document a déjà été importé pour ce projet.
+                </span>
+            {% endif %}
+        </div>
+
+        <div class="fr-radio-group">
+            <input type="radio"
+                   name="type"
+                   value="annexe"
+                   dsfr="dsfr"
+                   id="annexe"
+                   data-choose-document-type-target="radio"
+                   data-action="choose-document-type#setSelectedValue">
+            <label class="fr-label" for="annexe">
+                Annexe
+            </label>
+        </div>
+
+        <a class="fr-btn fr-btn--icon-left fr-icon-file-text-line fr-mt-4w"
+           data-href="{% url 'gsl_notification:upload-a-document' programmation_projet.id "type" %}"
+           data-choose-document-type-target="button">
+            Créer sur Turgot
+        </a>
+    </div>
+{% endblock notification_content %}
