+++ gsl_notification/utils.py
@@ -181,7 +181,7 @@ def get_s3_object(file_name):
 def return_document_as_a_dict(document: Arrete | ArreteSigne):
     return {
         "name": document.name,
-        "type": document.file_type,
+        "file_type": document.file_type,
         "size": document.size,
         "created_at": document.created_at,
         "created_by": document.created_by,
